Modularize SDFGrid into reusable modules
