+++ src/constants.js
@@ -1,0 +1,11 @@
+export const DENSE_W = 1024;
+export const DENSE_H = 1024;
+
+export const IDB_NAME    = 'SDFFieldDB';
+export const IDB_VERSION = 7;
+
+export const STORE_META  = 'meta';
+export const STORE_BASE  = 'base';        // Int16 SDF per-layer (kept)
+export const STORE_BASEZ = 'base_zero';   // Float32 zero template per schemaId
+export const STORE_LAYER = 'overlay_layers';
+export const STORE_LMETA = 'overlay_layers_meta';
